Add files via upload
Assignment -1
The issue in the code block of seaborn histplot is due to "ValueError: cannot re index from a duplicate axis". This error is due to the pandas 'concat' operation performed during combining the test and training data files. 
Because of this operation the indexes are duplicated from both the files leading to exception.
To resolve this simply we can re index the data frame again to create new indexes for the dataset.[ i.e data = data.reset_index()] would do the job.
and normalized histogram can also be plotted.